requirements.txt
requests>=2.32.3
python-dateutil>=2.9.0
libzim>=3.3.0
prometheus-client>=0.19.0
schedule>=1.3.0
PyYAML>=6.0.1
